# docs/openapi.yaml
openapi: 3.0.0
paths:
...
  /jobs/{jobId}/stream:
    get:
      operationId: JobsController_streamJob
      parameters:
        - name: jobId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_5
info:
  title: verificat.xyz API
  description: verificat.xyz fact-checking platform API
  version: 0.1.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
...