# services/api/src/jobs/jobs.controller.ts
import {
  Controller,
  Post,
  Get,
  Param,
  Req,
  Sse,
  Logger,
  HttpCode,
  HttpStatus,
  UnprocessableEntityException,
  Body,
} from '@nestjs/common';
import type { Request } from 'express';
import * as fs from 'fs';
import * as path from 'path';
import * as crypto from 'crypto';
import * as os from 'os';
import { Public } from '../auth/public.decorator';
import { JobsService } from './jobs.service';
import { JobsEventService } from './jobs-event.service';
import { Observable } from 'rxjs';

interface SseMessageEvent {
  data: string;
  id?: string;
  type?: string;
  retry?: number;
}

@Public()
@Controller('jobs')
export class JobsController {
  private readonly logger = new Logger(JobsController.name);

  constructor(
    private readonly jobsService: JobsService,
    private readonly jobsEventService: JobsEventService,
  ) {}

  @Post('verify-text')
  @HttpCode(HttpStatus.ACCEPTED)
  async verifyText(@Body('text') text: string) {
    if (!text || typeof text !== 'string' || text.trim().length === 0) {
      throw new UnprocessableEntityException('Text is required');
    }
    const claimId = crypto.randomUUID();
    return this.jobsService.enqueueVerification(claimId, text.trim());
  }

  @Post('upload')
  @HttpCode(HttpStatus.ACCEPTED)
  async uploadAudio(@Req() req: Request) {
    const rawChunks: Uint8Array[] = [];
    for await (const chunk of req) {
      rawChunks.push(chunk as Uint8Array);
    }
    if (rawChunks.length === 0) {
      throw new UnprocessableEntityException('No audio data received');
    }
    const audioBuffer = Buffer.concat(rawChunks);

    const tmpDir = path.join(os.tmpdir(), 'verificat-uploads');
    fs.mkdirSync(tmpDir, { recursive: true });
    const tmpFile = path.join(tmpDir, `${crypto.randomUUID()}.webm`);
    fs.writeFileSync(tmpFile, audioBuffer);

    const result = await this.jobsService.enqueueFromAudio(tmpFile);

    return { jobId: result.jobId, claimId: result.claimId };
  }

  @Get(':jobId/stream')
  @Sse()
  async streamJob(
    @Param('jobId') jobId: string,
  ): Promise<Observable<SseMessageEvent>> {
    this.logger.log(`SSE subscription for job ${jobId}`);
    return this.jobsEventService.watchJob(jobId);
  }
}